test(runtime): isolate recovery-state regressions
